+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH5HJ01DM2P4797SNGPPDNAX"
+        "01KH61GGX95NF42E9DPHVX88TY"
     )
     r"""The unique identifier of the data source"""