+++ packages/orq-rc/RELEASES.md
@@ -5509,3 +5509,13 @@ Based on:
 - [python v4.3.0-rc.51] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.51] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.51 - packages/orq-rc
+
+## 2026-02-11 09:48:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.709.4 (2.816.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.52] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.52] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.52 - packages/orq-rc